src/main/java/JavaGuru_2020_Java1/lesson_5/homeworks/level_5_6/ArrayUtil.java
package main.java.JavaGuru_2020_Java1.lesson_5.homeworks.level_5_6;

import java.util.Random;

public class ArrayUtil {

    public int[] createArray(int arrayLength) {

        int[] numbers = new int[arrayLength];
        return numbers;
    }

    public void fillArrayWithRandomNumbers(int[] array) {
        Random random = new Random();

        for (int i = 0; i < array.length; i++) {
            array[i] = random.nextInt(100);
        }
    }

    public void printArrayToConsole(int[] array) {
        for (int i = 0; i < array.length; i++) {
            System.out.println(array[i]);
        }
    }

    public int findMaxNumber(int[] array) {
        int maxNumber = array[0];

        for (int i = 0; i <array.length; i++) {
            if (array[i] > maxNumber) {
                maxNumber = array[i];
            }
        }
        return maxNumber;
    }

    public int findMinNumber(int[] array) {
        int minNumber = array[0];

        for (int i = 0; i < array.length; i++) {
            if (array[i] < minNumber) {
                minNumber = array[i];
            }
        }
        return minNumber;
    }
}
